move server ready check function for all e2e unit tests
