+++ pkg/server/wait_port_ready.go
@@ -1,0 +1,39 @@
+// Copyright 2020 The SQLFlow Authors. All rights reserved.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package server
+
+import (
+	"net"
+	"time"
+)
+
+func serverIsReady(addr string, timeout time.Duration) bool {
+	conn, err := net.DialTimeout("tcp", addr, timeout)
+	if err != nil {
+		return false
+	}
+	err = conn.Close()
+	return err == nil
+}
+
+// WaitPortReady will wait until the tcp address like ip:port can be connected from client.
+func WaitPortReady(addr string, timeout time.Duration) {
+	// Set default timeout to
+	if timeout == 0 {
+		timeout = time.Duration(1) * time.Second
+	}
+	for !serverIsReady(addr, timeout) {
+		time.Sleep(1 * time.Second)
+	}
+}
